+++ truffle/src/com.oracle.truffle.api.instrumentation.test/src/com/oracle/truffle/api/instrumentation/test/InstrumentationTest.java
@@ -40,6 +40,7 @@ import java.io.PipedOutputStream;
 import java.io.Reader;
 import java.nio.CharBuffer;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -86,7 +87,6 @@ import com.oracle.truffle.api.nodes.LanguageInfo;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
 import com.oracle.truffle.api.source.SourceSection;
-import java.util.Collections;
 
 public class InstrumentationTest extends AbstractInstrumentationTest {
 
@@ -155,6 +155,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
     @Test
     public void queryInstrumentsBeforeUseAndObtainSpecialService() throws Exception {
         engine = Engine.newBuilder().err(err).build();
+        context.enter();
         Runnable start = null;
         for (Instrument instr : engine.getInstruments().values()) {
             Runnable r = instr.lookup(Runnable.class);
@@ -164,6 +165,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
                 assertTrue("Now enabled: " + instr, isCreated(instr));
             }
         }
+        context.leave();
         assertNotNull("At least one Runnable found", start);
     }
 
@@ -217,8 +219,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         MultipleInstanceInstrument.onDisposeCounter = 0;
         MultipleInstanceInstrument.constructor = 0;
         Instrument instrument1 = engine.getInstruments().get("testMultipleInstruments");
-        engine.close();
-        engine = null;
+        cleanup();
         Assert.assertEquals(0, MultipleInstanceInstrument.constructor);
         Assert.assertEquals(0, MultipleInstanceInstrument.onCreateCounter);
         Assert.assertEquals(0, MultipleInstanceInstrument.onFinalizeCounter);
@@ -349,7 +350,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         }
 
         @Override
-        protected Context createContext(com.oracle.truffle.api.TruffleLanguage.Env env) {
+        protected InstrumentContext createContext(com.oracle.truffle.api.TruffleLanguage.Env env) {
             createContextCounter++;
             Instrumenter instrumenter = env.lookup(Instrumenter.class);
             Assert.assertNotNull("Instrumenter found", instrumenter);
@@ -1140,11 +1141,10 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         Assert.assertEquals("OutputToStdErrMoreOutputToStdErrEvenMoreOutputToStdErr", TestOutputConsumerArray.getErr());
 
         // instruments disabled
-        engine.close();
-        engine = null;
+        teardown();
+        setup();
         out.reset();
         err.reset();
-        engine = getEngine();
         context.eval(lines("PRINT(OUT, FinalOutputToStdOut)"));
         context.eval(lines("PRINT(ERR, FinalOutputToStdErr)"));
         Assert.assertEquals("FinalOutputToStdOut", out.toString());
@@ -1411,7 +1411,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         }
 
         assertEquals(0, service.onDisposeCalls);
-        context.close();
+        cleanup();
         assertEquals(0, service.onDisposeCalls);
         forked.getEngine().close();
         // test if all engines are disposed
@@ -1429,7 +1429,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         }
 
         @Override
-        protected Context createContext(com.oracle.truffle.api.TruffleLanguage.Env env) {
+        protected InstrumentContext createContext(com.oracle.truffle.api.TruffleLanguage.Env env) {
             instrumenter = env.lookup(Instrumenter.class);
             return super.createContext(env);
         }
@@ -1548,9 +1548,11 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         }
     }
 
-    private static void setupEngine(Source initSource, boolean runInitAfterExec) {
+    private void setupEngine(Source initSource, boolean runInitAfterExec) {
+        teardown();
         InstrumentationTestLanguage.envConfig.put("initSource", initSource);
         InstrumentationTestLanguage.envConfig.put("runInitAfterExec", runInitAfterExec);
+        setup();
     }
 
     @Test
@@ -1593,8 +1595,7 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         StringBuilder onCallsLogger = new StringBuilder();
         instrument.lookup(TestAccessInstruments.class).onCallsLogger = onCallsLogger;
         access.env.lookup(other, TestAccessInstrumentsOther.class).onCallsLogger = onCallsLogger;
-        engine.close();
-        engine = null;
+        teardown();
         assertEquals("FFDD", onCallsLogger.toString());
     }
 
@@ -1609,8 +1610,10 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         assertEquals("InstrumentTestLang", info.getName());
         assertEquals("2.0", info.getVersion());
 
+        context.enter();
         assertNotNull(access.env.lookup(info, SpecialService.class));
         assertEquals(InstrumentationTestLanguage.FILENAME_EXTENSION, access.env.lookup(info, SpecialService.class).fileExtension());
+        context.leave();
     }
 
     @Registration(id = "testAccessInstruments", name = "name", version = "version", services = TestAccessInstruments.class)
@@ -1670,22 +1673,10 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
 
     }
 
-    public class ReturnLanguageEnv {
-
-        public static final String KEY = "envReturner";
-
-        public TruffleLanguage.Env env;
-
-    }
-
     @Test
     public void testAccessInstrumentFromLanguage() {
-        ReturnLanguageEnv envReturner = new ReturnLanguageEnv();
-        InstrumentationTestLanguage.envConfig.put(ReturnLanguageEnv.KEY, envReturner);
-        context.eval(Source.create(InstrumentationTestLanguage.ID, ""));
-        assertNotNull(envReturner.env);
-
-        TruffleLanguage.Env env = envReturner.env;
+        context.initialize(InstrumentationTestLanguage.ID);
+        TruffleLanguage.Env env = InstrumentationTestLanguage.currentEnv();
         LanguageInfo langInfo = env.getLanguages().get(InstrumentationTestLanguage.ID);
         assertNotNull(langInfo);
         assertTrue(langInfo.getMimeTypes().contains(InstrumentationTestLanguage.MIME_TYPE));
@@ -1720,11 +1711,10 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         // Events during language initialization phase are included:
         TestLangInitialized.initializationEvents = true;
         TestLangInitialized service = instrument.lookup(TestLangInitialized.class);
+
         run("LOOP(2, STATEMENT())");
         assertEquals("[FunctionRootNode, false, StatementNode, false, ExpressionNode, false, ExpressionNode, false, FunctionRootNode, true, LoopNode, true, StatementNode, true, StatementNode, true]",
                         service.getEnteredNodes());
-        engine.close();
-        engine = null;
     }
 
     @Test
@@ -1738,8 +1728,6 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         TestLangInitialized service = instrument.lookup(TestLangInitialized.class);
         run("LOOP(2, STATEMENT())");
         assertEquals("[FunctionRootNode, true, LoopNode, true, StatementNode, true, StatementNode, true]", service.getEnteredNodes());
-        engine.close();
-        engine = null;
     }
 
     @Test
@@ -1755,8 +1743,6 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         assertEquals("[FunctionRootNode, false, StatementNode, false, ExpressionNode, false, ExpressionNode, false, " +
                         "FunctionRootNode, true, LoopNode, true, StatementNode, true, StatementNode, true, FunctionRootNode, true, StatementNode, true, ExpressionNode, true, ExpressionNode, true]",
                         service.getEnteredNodes());
-        engine.close();
-        engine = null;
     }
 
     @Test
@@ -1772,8 +1758,6 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         run("LOOP(2, STATEMENT())");
         assertEquals("[FunctionRootNode, true, LoopNode, true, StatementNode, true, StatementNode, true, FunctionRootNode, true, StatementNode, true, ExpressionNode, true, ExpressionNode, true]",
                         service.getEnteredNodes());
-        engine.close();
-        engine = null;
     }
 
     @Registration(id = "testLangInitialized", services = TestLangInitialized.class)
@@ -1817,8 +1801,6 @@ public class InstrumentationTest extends AbstractInstrumentationTest {
         assureEnabled(instrument);
         TestAllocation allocation = instrument.lookup(TestAllocation.class);
         run("LOOP(3, VARIABLE(a, 10))");
-        engine.close();
-        engine = null;
         assertEquals("[W 4 null, A 4 10, W 4 null, A 4 10, W 4 null, A 4 10]", allocation.getAllocations());
 
     }