+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/nodes/RootNode.java
@@ -150,8 +150,8 @@ public abstract class RootNode extends ExecutableNode {
     protected RootNode(TruffleLanguage<?> language, FrameDescriptor frameDescriptor) {
         super(language);
         CompilerAsserts.neverPartOfCompilation();
-        if (this.languageInfo != null) {
-            this.sourceVM = Node.ACCESSOR.engineSupport().getVMFromLanguageObject(this.languageInfo.getEngineObject());
+        if (this.language != null) {
+            this.sourceVM = Node.ACCESSOR.engineSupport().getVMFromLanguageObject(Node.ACCESSOR.languageSupport().getLanguageInfo(this.language).getEngineObject());
         } else {
             this.sourceVM = getCurrentVM();
         }
@@ -181,7 +181,7 @@ public abstract class RootNode extends ExecutableNode {
      * @since 0.27
      */
     public final <C, T extends TruffleLanguage<C>> C getCurrentContext(Class<T> languageClass) {
-        if (languageInfo == null) {
+        if (language == null) {
             return null;
         }
         return getLanguage(languageClass).getContextReference().get();